chore: move out grpc types into their own files
