+++ src/messages/responses/cache-dictionary-fetch.ts
@@ -6,6 +6,7 @@ import {
 } from './response-base';
 import {SdkError} from '../../errors/errors';
 import {TextDecoder} from 'util';
+import {_DictionaryFieldValuePair} from './grpc-response-types';
 
 const TEXT_DECODER = new TextDecoder();
 
@@ -30,15 +31,6 @@ const TEXT_DECODER = new TextDecoder();
  * ```
  */
 export abstract class Response extends ResponseBase {}
-
-export class _DictionaryFieldValuePair {
-  readonly field: Uint8Array;
-  readonly value: Uint8Array;
-  constructor({field, value}: {field: Uint8Array; value: Uint8Array}) {
-    this.field = field;
-    this.value = value;
-  }
-}
 
 class _Hit extends Response {
   private readonly items: _DictionaryFieldValuePair[];
