+++ src/messages/responses/grpc-response-types.ts
@@ -1,0 +1,42 @@
+/**
+ * This file exists as a layer between our javascript classes and the grpc generated
+ * ones. Its required since the types generated by grpc-web are different than those
+ * generated by grpc for node. Inside of the respective clients, we will map the grpc
+ * response types to these corresponding types. This allows us to share the response
+ * messages between the 2 sdks.
+ */
+
+export enum _ECacheResult {
+  Invalid = 0,
+  Ok = 1,
+  Hit = 2,
+  Miss = 3,
+}
+
+export class _DictionaryFieldValuePair {
+  readonly field: Uint8Array;
+  readonly value: Uint8Array;
+  constructor({field, value}: {field: Uint8Array; value: Uint8Array}) {
+    this.field = field;
+    this.value = value;
+  }
+}
+
+export class _DictionaryGetResponsePart {
+  readonly result: _ECacheResult;
+  readonly cacheBody: Uint8Array;
+
+  constructor(result: _ECacheResult, cacheBody: Uint8Array) {
+    this.result = result;
+    this.cacheBody = cacheBody;
+  }
+}
+
+export class _SortedSetGetScoreResponsePart {
+  readonly result: _ECacheResult;
+  readonly score: number;
+  constructor(result: _ECacheResult, score: number) {
+    this.result = result;
+    this.score = score;
+  }
+}
